Move telemetry resource tests to a single-threaded program
