+++ apollo-router/src/plugins/telemetry/resource.rs
@@ -47,7 +47,8 @@ impl ResourceDetector for EnvServiceNameDetector {
     }
 }
 
-pub(crate) trait ConfigResource {
+#[allow(missing_docs)] // only public-but-hidden for tests
+pub trait ConfigResource {
     fn service_name(&self) -> &Option<String>;
     fn service_namespace(&self) -> &Option<String>;
 
@@ -139,202 +140,4 @@ impl ResourceDetector for ConfigResourceDetector {
     }
 }
 
-#[cfg(test)]
-mod test {
-    use std::collections::BTreeMap;
-    use std::env;
-
-    use opentelemetry::Key;
-    use serial_test::serial;
-
-    use crate::plugins::telemetry::config::AttributeValue;
-    use crate::plugins::telemetry::resource::ConfigResource;
-
-    struct TestConfig {
-        service_name: Option<String>,
-        service_namespace: Option<String>,
-        resources: BTreeMap<String, AttributeValue>,
-    }
-    impl ConfigResource for TestConfig {
-        fn service_name(&self) -> &Option<String> {
-            &self.service_name
-        }
-        fn service_namespace(&self) -> &Option<String> {
-            &self.service_namespace
-        }
-        fn resource(&self) -> &BTreeMap<String, AttributeValue> {
-            &self.resources
-        }
-    }
-
-    // All of the tests in this module must execute serially wrt each other because they rely on
-    // env settings and one of the tests modifies the env for the duration of the test. We enforce
-    // this with the #[serial] derive.
-    #[test]
-    #[serial]
-    fn test_empty() {
-        let test_config = TestConfig {
-            service_name: None,
-            service_namespace: None,
-            resources: Default::default(),
-        };
-        let resource = test_config.to_resource();
-        assert!(resource
-            .get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into())
-            .unwrap()
-            .as_str()
-            .starts_with("unknown_service:apollo_router"));
-        assert!(resource
-            .get(opentelemetry_semantic_conventions::resource::SERVICE_NAMESPACE.into())
-            .is_none());
-        assert_eq!(
-            resource.get(opentelemetry_semantic_conventions::resource::SERVICE_VERSION.into()),
-            Some(std::env!("CARGO_PKG_VERSION").into())
-        );
-
-        assert!(resource
-            .get(opentelemetry_semantic_conventions::resource::PROCESS_EXECUTABLE_NAME.into())
-            .expect("expected excutable name")
-            .as_str()
-            .contains("apollo"));
-    }
-
-    #[test]
-    #[serial]
-    fn test_config_resources() {
-        let test_config = TestConfig {
-            service_name: None,
-            service_namespace: None,
-            resources: BTreeMap::from_iter(vec![
-                (
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAME.to_string(),
-                    AttributeValue::String("override-service-name".to_string()),
-                ),
-                (
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAMESPACE.to_string(),
-                    AttributeValue::String("override-namespace".to_string()),
-                ),
-                (
-                    "extra-key".to_string(),
-                    AttributeValue::String("extra-value".to_string()),
-                ),
-            ]),
-        };
-        let resource = test_config.to_resource();
-        assert_eq!(
-            resource.get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into()),
-            Some("override-service-name".into())
-        );
-        assert_eq!(
-            resource.get(opentelemetry_semantic_conventions::resource::SERVICE_NAMESPACE.into()),
-            Some("override-namespace".into())
-        );
-        assert_eq!(
-            resource.get(Key::from_static_str("extra-key")),
-            Some("extra-value".into())
-        );
-    }
-
-    #[test]
-    #[serial]
-    fn test_service_name_service_namespace() {
-        let test_config = TestConfig {
-            service_name: Some("override-service-name".to_string()),
-            service_namespace: Some("override-namespace".to_string()),
-            resources: BTreeMap::new(),
-        };
-        let resource = test_config.to_resource();
-        assert_eq!(
-            resource.get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into()),
-            Some("override-service-name".into())
-        );
-        assert_eq!(
-            resource.get(opentelemetry_semantic_conventions::resource::SERVICE_NAMESPACE.into()),
-            Some("override-namespace".into())
-        );
-    }
-
-    #[test]
-    #[serial]
-    fn test_service_name_override() {
-        // Order of precedence
-        // OTEL_SERVICE_NAME env
-        // OTEL_RESOURCE_ATTRIBUTES env
-        // config service_name
-        // config resources
-        // unknown_service:executable_name
-        // unknown_service (Untested as it can't happen)
-
-        assert!(TestConfig {
-            service_name: None,
-            service_namespace: None,
-            resources: Default::default(),
-        }
-        .to_resource()
-        .get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into())
-        .unwrap()
-        .as_str()
-        .starts_with("unknown_service:apollo_router"));
-
-        assert_eq!(
-            TestConfig {
-                service_name: None,
-                service_namespace: None,
-                resources: BTreeMap::from_iter(vec![(
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAME.to_string(),
-                    AttributeValue::String("yaml-resource".to_string()),
-                )]),
-            }
-            .to_resource()
-            .get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into()),
-            Some("yaml-resource".into())
-        );
-
-        assert_eq!(
-            TestConfig {
-                service_name: Some("yaml-service-name".to_string()),
-                service_namespace: None,
-                resources: BTreeMap::from_iter(vec![(
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAME.to_string(),
-                    AttributeValue::String("yaml-resource".to_string()),
-                )]),
-            }
-            .to_resource()
-            .get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into()),
-            Some("yaml-service-name".into())
-        );
-
-        env::set_var("OTEL_RESOURCE_ATTRIBUTES", "service.name=env-resource");
-        assert_eq!(
-            TestConfig {
-                service_name: Some("yaml-service-name".to_string()),
-                service_namespace: None,
-                resources: BTreeMap::from_iter(vec![(
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAME.to_string(),
-                    AttributeValue::String("yaml-resource".to_string()),
-                )]),
-            }
-            .to_resource()
-            .get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into()),
-            Some("env-resource".into())
-        );
-
-        env::set_var("OTEL_SERVICE_NAME", "env-service-name");
-        assert_eq!(
-            TestConfig {
-                service_name: Some("yaml-service-name".to_string()),
-                service_namespace: None,
-                resources: BTreeMap::from_iter(vec![(
-                    opentelemetry_semantic_conventions::resource::SERVICE_NAME.to_string(),
-                    AttributeValue::String("yaml-resource".to_string()),
-                )]),
-            }
-            .to_resource()
-            .get(opentelemetry_semantic_conventions::resource::SERVICE_NAME.into()),
-            Some("env-service-name".into())
-        );
-
-        env::remove_var("OTEL_SERVICE_NAME");
-        env::remove_var("OTEL_RESOURCE_ATTRIBUTES");
-    }
-}
+// Tests in apollo-router/tests/telemetry_resource_tests.rs
